Move management of OpenGL context to CC3OpenGL.
Move management of OpenGL context to CC3OpenGL from CC3GLView.
Add CC3OpenGL deleteGL to delete all instances of CC3OpenGL.
Add CC3GLView.h file to consolidate platform-specific view files.
Remove CC3GLBackgrounder. Move backgrounder property to CC3Scene.
Remove CC3OpenGL _isRenderingContext iVar.
Replace several #include directives with #import.

+++ cocos3d/cocos3d/Shaders/CC3Shaders.m
@@ -1096,7 +1096,7 @@ static CC3Cache* _shaderSourceCodeCache = nil;
 
 -(NSString*) sourceCodeString { return _sourceCodeString; }
 
--(GLuint) lineCount { return _sourceCodeString.lineCount; }
+-(GLuint) lineCount { return (GLuint)_sourceCodeString.lineCount; }
 
 -(void) appendSourceCodeString: (NSString*) srcCode { [_sourceCodeString appendString: srcCode]; }
 
@@ -1142,7 +1142,7 @@ static CC3Cache* _shaderSourceCodeCache = nil;
 	return srcCodeString;
 }
 
--(GLuint) lineCount { return _sourceCodeLines.count; }
+-(GLuint) lineCount { return (GLuint)_sourceCodeLines.count; }
 
 -(void) appendSourceCodeString: (NSString*) srcCode {
 	[_sourceCodeLines addObject: [NSData dataWithBytes: srcCode.UTF8String
@@ -1159,7 +1159,7 @@ static CC3Cache* _shaderSourceCodeCache = nil;
 
 -(void) accumulateSourceCompilationStringCountWithVisitor: (CC3ShaderSourceCodeCompilationStringCountVisitor*) visitor {
 	if ( [visitor hasAlreadyVisited: self] ) return;
-	[visitor addSourceCompilationStringCount: _sourceCodeLines.count];
+	[visitor addSourceCompilationStringCount: (GLuint)_sourceCodeLines.count];
 }
